fix(magnification): swipe up gesture in app does not work sometimes when fullscreen magnification activated
Previously, we added the MagnificationInfoChangedCallback to the gesture handler. When the gesture handler is notified the magnification deactivation, we can reset the gesture handler states. However, when fullscreen magnification activated and always-on-magnification is not enabled, the swipe up gesture would trigger the context changed event, then the magnification would be deactivated, and the gesture handler states reset. It would cause the gesture handler does not delegate the complete gesture events, then the system would not detect the complete swipe up gesture and perform the correct ui logic.

Therefore, when gesture handler is notified that the magnification is deactivated, instead reseting all states, we only cancel the shortcut triggered flag. By doing this, we can ensure that future taps will not activate the magnifier unexpectedly, and the gesture handler will process the remaining events properly.

Bug: 298023290
Test: manually
Change-Id: I8643ff159216e7511d42e1c63e7300bd1133b231

Former-commit-id: cd5748602509f2eab69b8d90790395741608a9bf