Request transition when disposing TF organizer
To make sure the changes are collected. Otherwise if the activity
in the TaskFragment is the top app, the previous playing transition
won't be aware of the next resuming activity should be visible.

For example, if there is an opening transition that shows B and
hides A, but A is died suddenly and disposing its TaskFragment,
then B should show again. With the requested CLOSE transition,
the playing transition will be able to merge the transition and
update the surface visibility for A (TO_FRONT) in finishTransaction.

Bug: 288565944
Bug: 287895832
Bug: 297315249
Test: Launch an app which supports embedded activity.
      E.g. Settings with window extensions and large_screen_opt.
      Launch Settings and open SubSettings. Press home key.
      Launch Settings again and enter the command immediately
        adb shell kill -9 $(adb shell pidof com.android.settings)
      Home should still be visible.
Change-Id: I36703c8d54d3b9539ca7987ba62db9dd235b3909

Former-commit-id: 44dbfdb975ca07f4b23e639ca70b64f3558cdedc