Set cursor position in handwriting initiation
When handwriting is initiated for an unfocused EditText, if the stylus
down point was inside the EditText's bounds, then the EditText will
automatically set its cursor position nearest to the stylus down point
when it gains focus. If the stylus down point was outside the EditText's
bounds (within the extended handwriting bounds), then we must calculate
and set the cursor position manually.

Bug: 288461219
Test: atest HandwritingInitiatorTest
Change-Id: I62593f1ff01a7ec39c2eb6e1352d63a7f8b79f88

Former-commit-id: 88f4cc5d3233d0b9491de4103f9f4e084cd1f341