Avoid sending IME insets position in old rotation to client
IME token surface is hidden by reusing AsyncRotationController to
avoid showing intermediate state when running app transition with
rotation change. But it is not the accepted window type for
async rotation, so it can be excluded from freezing position.

This prevents the case if the client InsetsAnimationControlImpl
is late to call updateSurfacePosition by the old rotation after
server side updated to new position.

Bug: 297346129
Test: atest ShowImeWhileEnteringOverviewTest# \
      imeLayerIsVisibleAndAssociatedWithAppWidow
Change-Id: I6a79a631490c250391102e4dec5f3335a7ed0fea

Former-commit-id: 534cc38c0d1468475e563085fc580171a3ce1bc5