Do not hide IME when waiting for drawn by toggling screen
This removed outdated logic of [1].
It was needed because the IME container was put on top of display.
[1]: Ife93bdfde8ba2914930497356c0e16ebd629c507

Since DisplayArea is introduced, the corresponding ImePlaceholder
will never have higher z-order then AOD unless itself is the IME
target. Also since insets leash is introduced, the leash surface
will be hidden if the control target is not visible. So the logic
is no longer needed.

Otherwise if WindowState#hide is called for IME, it will clear
visibility flag and cause WindowState#isGoneForLayout to return
true. Then IME won't receive new size in time when the display is
resizing with turning off/on screen.

Bug: 289995722
Test: Fold/unfold device when "Stay unlocked on fold" is enabled
      and the current app is showing IME. The IME won't show a
      few frames in old size.
Change-Id: I697e1ec679be37388eefb0fa265d9e9a084458cf

Former-commit-id: 8c6f03faa69792582b33dc4a7c8ddcc4f9311eb0