Partial Screen Sharing: fix last item in app selector being empty
The bug was related to the override of shouldShowContentPreview() in
MediaProjectionAppSelectorActivity.
This method was being used internally in ChooserActivity not only to
determine whether the content preview should be shown, but also
whether a "service target" row should be shown or not. By default that
method checks whether the Intent action is ACTION_SEND.
MediaProjectionAppSelectorActivity overrides the method to always return
true. Then there is code in ChooserListAdapter that also checks for
the intent action to determine the service target count. This was
causing an inconsistency between ChooserActivity and ChooserListAdapter,
making it do a wrong calculation for the list item indexes.

Change-Id: I3dee38909035c44816d9ae1d06b17eec6266107e
Test:  atest FrameworksCoreTests:ChooserListAdapterTest
Test:  atest FrameworksCoreTests:ChooserActivityTest
Fixes: 295877430
Fixes: 299347412

Former-commit-id: 7a5061612f69a6c61f26403c769ac904076a59ea