[3/n] Optimize user aspect ratio button heuristic
If `updateCompatInfo` is called while the user aspect ratio button is
showing, the buton will be removed as we no longer consider the layout
to have a button as it has already been shown.

Update `getHasUserAspectRatioSettingsButton` so the layout is consider
to have a button for the duration the button is visible and only once
the button has been shown and timed out, will the layout be considered
to not have a button.

Test: atest WMShellUnitTests:UserAspectRatioSettingsWindowManagerTest
Bug: 301323072
Bug: 289356588
Change-Id: Ie4be836d52b74e285a501a799094022be24870fe

Former-commit-id: e17d87347d611e52d022bbf246c933a2222a1ed7