Remove FrameRate Enum to allow all possible frame rate values
Previously we restricted the frame rate override of game intervention to
a certain set of values, which is now not scalable and not compatible to
all devices.

In addition, SurfaceFlinger is able to choose the proper
frame rate to run at, there is no longer a need to impose such
restriction to available frame rates.

Bug: 296988493
Bug: 253178812

Tested on Pixel 7 Pro userdebug
Test: atest GameManagerServiceTests
Test: atest GameManagerTests
Test: atest CtsFrameRateOverrideTestCases

Change-Id: I2e2e18a9ec0d1a2fd505464a1297e9b7f9874d1c

Former-commit-id: b3e07d83c8f533fe5efc956f7eca34d8bc3b164a