Always allow UDFPS touches if alt bouncer is visible
QS can be mid-expansion but validly require UDFPS auth when
notifications are not allowed to be shown on the lockscreen.
When QS is in mid-expansion, we request/show the AlternateBouncer
so the user has an opportunity to use UDFPS to see their notifications.

Test: atest UdfpsControllerTest
Test: manual (don't show notifications on LS,
can still use UDFPS when requested to auth
when pulling down the shade on LS)
Fixes: 301037136

Change-Id: I5c91c4df394f8c6788077dd1ba60fe3123993d54

Former-commit-id: ccfdf94b931250a90e993207b9c000f2d55b7b67