fix(statusbar): notification icons become too small when smaller font size
In StatusBarIconView, when calculating the icon scale for notification icons, we should first compute the scale to make the raw drawable same size when icon view size is mOriginalStatusBarIconSize. Therefore, when smaller font size already makes the drawable constrained further than mOriginalStatusBarIconSize does, we should compute the scale to enlarge the constrained image to the size it would be when mOriginalStatusBarIconSize. Then, we could do the next steps to scale the scaled image to fit mNewStatusBarIconSize.

Bug: 296026932
Test: manually - attach screenshots in bug
      atest systemui/statusbar/phone/
Change-Id: Id8aa77ebac631377cc2ed623d8c4dedad59c27de

Former-commit-id: 9641493d7005dd374d5b14f034b7ce46887a1e6b