Truncate user data to a limit of 500 characters
Fix vulnerability that allows creating users with no restrictions. This is done by creating an intent to create a user and putting extras that are too long to be serialized. It causes IOException and the restrictions are not written in the file.

By truncating the string values when writing them to the file, we ensure that the exception does not happen and it can be recorded correctly.

Bug: 293602317
Change-Id: Ia71477601d036a3ca55e73cdc9698ae268a30f20
Test: install app provided in the bug, open app and click add. Check logcat to see there is no more IOException. Reboot the device by either opening User details page or running adb shell dumpsys user | grep -A12 heen and see that the restrictions are in place.
Merged-In: I633dc10974a64ef2abd07e67ff2d209847129989

Former-commit-id: 46caac641941f2e8865a8d53400f959b3bd98d88