Fix getWallpaperSafeLocked
The first time this method is called for a system-only wallpaper, if the
lock wallpaper is not loaded yet, it will not be aware of the lock
wallpaper and will think that the system wallpaper is sys+lock.

This happens after a reboot + user switch when the secondary user have
two different wallpapers.

This may fix other cases where the lock screen wallpaper is set to same
as home screen

Bug: 300537586
Test: atest WallpaperManagerTest
Test: manually check that the bug is fixed
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:58e3cfe430ed60bd4bb56117824d6736831562ef)
Merged-In: Ib37f83233deac7198997910991ffcce75b357aae
Change-Id: Ib37f83233deac7198997910991ffcce75b357aae

Former-commit-id: be88ec9e5d18d9b7f72807d06cc5195f253d7aa5