Add landscape layouts for keyguard sim pin and puk
When lockscreen has rotation allowed and a sim card that requires
pin to unlock is inserted it is not possible to unlock the sim pin.
The reason for this is that the current layout is not made for
a keyguard in landscape, the digit pad is not visible for example.

Add a landscape layout that takes lockscreen rotation into account.

Flag: NA
Bug: 325046250
Test: atest KeyguardSimPinViewControllerTest KeyguardSimPukViewControllerTest
Change-Id: I49faf4273344bf4abfa078930b8c4e74c53fc4b8

Former-commit-id: 9c62a2cfe666df892ec2fc4c5f66421a5225a3c1