Fix IME show delay when having a RemoteInsetsControlTarget
With the refactor in [1] and [2], the IME did not appear immediately when having a controlTarget that is not the app, but with 1-2s delay.
The IME leash is only dispatched when it is successfully drawn. To set that state, it has to be serverVisible and clientVisible. Before, we did not set the mClientVisible in the ImeInsetsSourceProvider.
By setting the clientVisibility directly (based on the requestedVisibility of the app for the IME), there will be no more delay.

[1]: I8e3a74ee579f085cb582040fdba725e7a63d6b85
[2]: I5aeb3ccfbfe7d4cd0b7b5f0e0400769e65bb70a4

Test: manual: Open Bubbles and try to show the IME. Verify that there is no noticeable delay between focussing and when the IME appears.
Flag: android.view.inputmethod.refactor_insets_controller
Fix: 362478040
Change-Id: I2d37f74a16205f6a75a2fa035b91641fc5246077

Former-commit-id: 798f9846e5dc64d6ff5a0f3b0eb888096ba04801