Associate activity config before creating application
In case when cold launching an activity with letterboxed bounds, the
app stores the size or configuration in its Application#onCreate,
which could be still from relative to entire display. And then
Activity#onCreate may initialize with letterboxed configuration,
which may cause UI issue if it uses the stored size from Application.

So this change makes the configuration consistent in both
Application#onCreate and Activity#onCreate.

Bug: 345582967
Flag: EXEMPT bugfix
Test: atest RootWindowContainerTests#testAttachApplication
Change-Id: I7604671446901c87cb1440022327b562dcb4baf6

Former-commit-id: bebd32ea77ec7b4b9bf6acec4540f1ff68711e1a