Rework SDK_MINOR_INT -> SDK_INT_FULL
Change how minor release codes are encoded in Build.java. Instead of
a new int representing the minor version, which always needs to be
combined with SDK_INT, change SDK_MINOR_INT to become SDK_INT_FULL which
encodes both major and minor versions. This reduces the risk of
confusing the integers, simplifies version checks and is better for Lint
and the docs.

Bug: 350458259
Test: atest 'CtsOsTestCases:BuildTest#testSdkIntFull'
Flag: android.sdk.major_minor_versioning_scheme
Change-Id: I4373e846f572aedd2f48af17c8d60aea80c42b2e

Former-commit-id: 31db47319937f234c8eae3d6c1b6ab6bfc2b9ed3