Cleaning up quick settings flag & removing redundancies
With a11y_qs_enabled being removed,
we are now able to remove a large amount of duplicate code regarding shortcuts,
and centralize most of the logic in manager & shortcutUtils.

This change also adds a new fxn to ShortcutUtils that will allow for further cleanup in Settings.

Test: atest com.android.settings.accessibility com.android.internal.accessibility
Bug: 367414968
Flag: EXEMPT flag cleanup

Change-Id: I7e744df9cf8a6c764448d1bdd4d45973c2b5edcb

Former-commit-id: 4cab65422a4b2cf5ec6f455f81c4f97da28892e1