PagerDots animation fix
This fixes the animation when paging (that was completely not working)
as the animated value node was in a branch where it had a fixed value,
so that had to be hoisted. After that, "simplify" the drawing logic at
the cost of two drawCircle (for inactive circles behind the active
rounded rect).

Test: manual
Fixes: 369156081
Flag: com.android.systemui.qs_ui_refactor_compose_fragment

Change-Id: I751ad971dc186afd9539ed6646a789e8e75374fc

Former-commit-id: db468ddf06a62e6119594adf1df58b88d96d4070