Make RecordIssueTile's ServiceConnections not injected, but created.
This fixes the issue of sometimes having the dagger injection giving the
same instance in the tile and the service, and sometimes not. We can
avoid this by ensuring the service connections are always different.

Also, by Binding to IssueRecordingService, we can avoid "leaking" the
traceurConnection which would otherwise need to be created / destroyed
with every notification action.

Bug: 364824477
Test: Verified locally that System Tracing via the Record Issue Tile
works great in Headless System User Mode.
Flag: EXEMPT bug fix

Change-Id: I076f7a9c058fbc725296a3c2bb34cbc75675d30c

Former-commit-id: 70b54e38d928b308bea19fdd9bf5d55d1125ca1a