Add CancellationSignal to AppFunctionManager API.
Marking the existing functions as deprecated and not removing them because we have a lot of code referencing these previous APIs and we will remove these once those references have been updated. Since the usage is in G3 we can't do it right away.

The sidecar update and logic to wrap cancellation signal with timeout and other checks will be added in later cls.

Change-Id: I48007082315a23b395d135b7b2130a05df82e9a4
Flag: android.app.appfunctions.flags.enable_app_function_manager
Test: atest CtsAppFunctionTestCases
Bug: 360864791

Former-commit-id: 6fb904965e77ad17a1b95a1b976902a5d0f3eaa4