Fix mediaOffset for QQS flexiglass
The problem is that the tiles in the QQS are centered inside of a
container, that has an offset on the bottom. This change centers
MediaCarousel in the QS and applies a corresponding offset to it. More
details and a picture is in the bug.

Flag: com.android.systemui.scene_container
Fixes: 360130533
Test: manual on foldable: observe media carousel in all possible configurations
Change-Id: I58187f9f7219a1a7f9c3f71c818fe0e51d22cebd

Former-commit-id: 90b93d50414eb4ad923712ca40bb6d466b80caad