Expand the a11y focus bounds if they're too small to be visible.
Attempting to draw the focus drawable with bounds that are smaller
than 2 times the focus stroke size results in it being invisible.
Normally Views shouldn't be this thin in the first place, but this
change helps ensure the focus is visible when drawn over those edge
case ultra-thin Views.

See the bug comment 10 for screenshots of the fix.

Flag: android.view.accessibility.focus_rect_min_size
Test: atest FrameworksCoreTests:ViewRootImplTest
Test: Open problem app from bug, observe focus rect is now visible
Bug: 368667566
Change-Id: I5261cbd980d2a11a28a755756a4104ac2a349870

Former-commit-id: 0b1a4fbfc8e374aaf4c8682332736bdf1428d69f