Adding MSDL haptic feedback when pulling down the shade.
This happens for both single and dual-shade scenarios by considering all
transitions from the "Gone" scene towards other scenes and towards the
overlays. A SceneContainerHapticsViewModel has been created for this
purpose. To co-exist with the performHapticFeedback API, the view-model
can be created with a reference to a View (expected to be removed once
the MSDL flag is set to true).

Test: atest SceneContainerHapticsViewModelTest
Flag: com.android.systemui.msdl_feedback
Bug: 352771261

Change-Id: Ice2a5b1b6fcb72ec7f5d9892dae86cc96e5b13d5

Former-commit-id: 6e7adcf3c5ffa38e0206ff56cc0ae65d6131d4ac