Log widget tap and scroll events
This change adds InteractionHandler APIs for logging tap and scroll
events, and connects them to RemoteViews.

The tap and scroll events includes the ID of the clicked item.

This change also adds a RemoteViews.setUsageEventTag API to be used to
set a tag to be used for event logging. The
android.R.id.remoteViewsMetricsTag public resource is removed.

Bug: 364655296
API-Coverage-Bug: 404344893
Test: AppWidgetEventsTest#interactionLogger_*
Flag: android.appwidget.flags.engagement_metrics
Change-Id: I90de107a93f9d3183a09656242cb3a7e2ef71f83

Former-commit-id: 4f871ab5855bd9a0b5add21a44ec5075d4fc1fb7