[flexiglass] Communal scene navigates to shade and bouncer.
Defines navigation user actions between the communal scene and the shade
scene as well as the bouncer scene. Also adds navigation back to Gone if
the user swipes right or up when the device is unlocked.

Still missing:
- The transition animations are missing and will be added in the next CL
- The shade scene returns to lockscreen instead of communal

Bug: 323068793
Bug: 339449476
Test: manually verified navigation. Swipe up for bouncer and swipe
right for lockscreen - when locked. Could not test when unlocked as I
couldn't figure out a way to unlock while on GH.  Tested that swipe down goes to shade when locked and
when unlocked.
Test: added unit test for new user action view-model
Flag: com.android.systemui.scene_container

Change-Id: I2966348858911eb69ef7fbb0d81bb90c11f4360b

Former-commit-id: 2bd406ba9e66bcc4ed1ac3c4ad7408aa37d0c35d