Remove redundant caller check.
Result of this call is not used. And there is no reason to call it for
side effects either: this check can only be performed when factoryRestet
is not null, and this only happens when invoking from
wipeDataWithReason, which already has identical call, so any permission
checks are already performed.
Without that check the code no longer needs valid calling identity, so
the method can again work as a single binderWithCleanCallingIdentity
block.

Also changed system user precondition check from authorization to state
assertion for attempts to remove the system user to make it match the
javadoc.

Fixes: 286212112
Change-Id: I3c604d27e0ab0d2f795feda44716c067c5816865
Test: btest a.d.c.WipeDataTest
Flag: EXEMPT bugfix

Former-commit-id: 38186761f903b9d4f8a0a84a574271994541735c