Restrict use of framework-annotations
This filegroup target should only be used when the raw srcs are needed.
Module targets can use `framework-annotations-lib` as a libs reference,
if necessary. Otherwise, the annotations are bundled directly with the
framework.

This refactor is in anticipation of bundling keep rules with the
framework-annotations-lib target, instead of defining global keep rules.

Bug: 275175207
Test: m
Flag: EXEMPT refactor
Change-Id: Id3d224ae725981151cdfe3d877d4ca152285842c

Former-commit-id: e8d6edc6e714826c8136bd9600e0442793da97a0