Manual network selection handling when request satellite enable
If the device is in  manual network selection mode and the user starts a satellite session then
the underlying modem will not be able to connect to the satellite network (Skylo) as Skylo is a
different network.

When a terminal attempts to use satellite communication, there is a high probability that there
will be no available networks. Therefore, the processing continued with the rest of the satellite
operations without waiting for a successful or failed response from the Automatic network
selection req.

Bug : 374301122
Flag: EXEMPT strings addition
Fix: request from google by 'Samsung features for carrier roaming NB-IoT NTN', req 10
Test: trigger RequestSatelliteEnable with manual network selection state
Change-Id: I790111725536f215856dfd4ff4f5d3d4f1df9d5b

Former-commit-id: 4968679825c9069502b8529a6008feb31ac5531f