[1/3] Refactor animator creation to prepare for the new spring.
For return animations, in some scenarios we will be taking over
transitions in which the outgoing window has momentum (e.g. gesture nav,
predictive back). In these cases, using the default interpolators can
cause a jarring stagger between the tracking motion and the non-tracking
portion. Instead we will use a spring which can keep the momentum while
maintaining a similar duration and curve as the interpolators.

Bug: 323863002
Flag: EXEMPT refactor only
Test: atest TransitionAnimatorTest
Change-Id: I29ce76fd6a739144ff4945d3c9a6da0b1a0fb1db

Former-commit-id: a2930f6bbecc8f85c3cbc1805e16231ccc76ec98