Fix flashing when folding to AOD
LightRevealScrimInteractor was launching
scrim animation when folding to AOD
which lasts 500ms. This was causing the content
behind the notifications shade to be sometimes
visible momentarily right after folding.

Updated the logic to immediately set
scrim progress to '0' when folding to AOD.
This allows to update the scrim 500ms earlier,
which reduces the chances to see the content
behind the shade after folding.

Fixes: 357169978
Test: atest LightRevealScrimInteractorTest
Test: fold multiple times manually
Flag: EXEMPT bugfix
Change-Id: I0a95baa245da58f50449ccfa7692e80252548c91

Former-commit-id: 749725678e2b2e8d980993647350615a922e5268