Register VcnManager with VcnFrameworkInitializer
When VCN is moved out of framework.jar, SystemServiceRegistry
will not be able to register VCN service via IVcnManagementService
anymore because the aidl interface will not be accessible. To
solve this issue, this CL follows
go/android-api-guidelines#the-moduleframeworkinitializer-pattern
to create an FrameworkInitializer for service registration.
Followup CLs will expose this class as @SystemApi.

Bug: 369703242
Test: atest CtsVcnTestCases
Flag: EXEMPT code refactoring; no functional change
Change-Id: Ic9668a60f195f0a21de5114c1bb0bc38b72ee4d5

Former-commit-id: e4fff766f54bdeee46f1e2fef88c0005afc2cba9