Create ManagerRecord#notifySessionReleased
This change is part of a series of non-functional refactors to
encapsulate binder operations in the record that corresponds to the
target binder object. The objective is to improve the logging associated
with binder failures by adding information that's available in the
record. For example: Package names and ids.

Bug: 360129098
Test: atest CtsMediaBetterTogetherTestCases MediaRouter2HostSideTest
Flag: EXEMPT refactor
Change-Id: I00352a3974b6f60cbf0e75b6196375a53a7c2ed9

Former-commit-id: 945a81b5c0d22a3d698af5a5e6fcec8716fde8dd