Add DeviceEffectsDiff and PolicyDiff to RuleDiff
In the ZenModeDiff.RuleDiff, instead of comparing the ZenPolicy and
ZenDeviceEffects objects using the Object equals, we use the newly added
DeviceEffectsDiff and PolicyDiff objects. This allows us to actually see
the parts of the policies that are changing.

Also orders the string versions of these diffs by when the subfields
were added which should make them easier to compare.

Bug: 319241807
Test: atest ZenModeDiffTest
Flag: android.app.modes_api
Change-Id: I3139b437f3e5ae431fc13a595af59b37262ac840

Former-commit-id: a6aacbdb97dfa7c1d88ab75785355a39dbbe9893