make silence emergency nofitication broadcast protected
A user reported the Emergency Calling Wifi Notification kept
popping up and the Do Not Ask Again button was not working.

Upon inspection, the ActivityManager was blocking the broadcast
action because the broadcast was not protected.  Also, the
notification should be dismissed when clicked.

Flag: com.android.internal.telephony.flags.stop_spamming_emergency_notification
Bug: 368690328
Test: manual
Change-Id: Id1d2f8334ce4cd3e9a08d1c9792b171201532d37

Former-commit-id: 8840009964d31942b13ea30e4e1e3fca79f873d8