Correct an ApplicationSharedMemory unit test
The test `ApplicationSharedMemory.canRead()` verifies that the network
time can be read from shared memory.  The read will throw if the time
is not initialized, either because there is no time source or because
the feature that writes to shared memory is disabled.

This change accepts the particular `DateTimeException`.  Note that an
exception is thrown only if the memory was actually read, which means
the test passed.

Flag: EXEMPT test-only
Bug: 365575551
Test: atest
 * ApplicationSharedMemoryTest32
Change-Id: I455cb5b6e8b8a04557929495dfda8979f9da8e89

Former-commit-id: aa381d3b4800b5e004afb3d7dfbb5819ca175d6c