Handle apexd failures
Rethrowing RemoteException causes PackageManager crash. This happens
when a device shutting down (init kills apexd while it's being used by
PM).

Instead, apexd logs an error and returns a reasonable default value
(null or empty), except wrapping it in a PackageManagerException in a
few methods, which is handled by PM.

Bug: 370264486
Change-Id: Ie8aed22f16599f7e1769df2abd4a1e13013d9105
Test: ApexManagerTest

Former-commit-id: 5557dd75eeba18837b0585208ce1581a25502e0d