Reapply "Don't install a pointer input when there are no user actions"
This is a reland of ag/29669127, which was broken (and reverted in
ag/29690066) because SwipeToSceneNode was not properly handling the
update of its draggableHandler.

This reland fixes the issue by introducing a SwipeToSceneRootNode, which
recreates a new delegate node whenever draggableHandler has changed.
This approach was suggested by the Compose team in
http://shortn/_MUgPp4TRmY. A test covering this case was added in
SwipeToSceneTest.swipeToSceneSupportsUpdates().

Bug: 370913106
Test: atest SwipeToSceneTest
Flag: com.android.systemui.scene_container
Change-Id: Ifc1b88a813b6de153028db64aac6cd1103c94a66

Former-commit-id: 04a358e0836818274b4cb868cabe45d049c2a282