Allow VDM to access default device camera
By default, allow a virtual device to access the default
device camera and add an API for the virtual device owner
to explicitly disallow this during the creation of the virtual
device. Also, clean up the old stream_camera flag, as this
change is guarded by a new flag and a targetSdk check.

Test: atest VirtualDeviceParamsTest
Bug: 371173368
Flag: android.companion.virtualdevice.flags.default_device_camera_access_policy
Change-Id: Iedf68beb2a9f3003f675e125605facf8837dcb1b

Former-commit-id: 604bf23d0e7af974ba1bbb44b021ab614a0594a1