Ensure screenshot dismissal velocity is never 0
It's possible to get a value of 0 passed in to the dismissal animation,
which isn't handled correctly when attempting to set a minimum velocity.
This change fixes the minimum velocity calculation so it is correct even
if the input is 0f.

Bug: 362859928
Fix: 362859928
Test: manual. invoke screenshots, move screenshot UI and let it return
to its original location, then move it out again and move pointer
straight up and release. screenshot should dismiss normally without
getting 'stuck'.
Flag: EXEMPT trivial fix

Change-Id: Idc963cf322d11be007b5e9167e783c7a2661a03b

Former-commit-id: de7b75b4e0e964080b23e0e2e1ed65ecc9d450ec