Exit desktop immersive if applicable on other transitions
When in desktop immersive, there are a few scenarios in which the
immersive state should be broken and the immersive task should return to
the maximized bounds, for example:
- The immersive task is minimized (button or otherwise)
- Another task is launched on top
- Another task is brought to front
- We re-enter the desktop that had an immersive app with a new app in
  front (drag or handle menu entries)
- etc...

This change notifies the immersive handler during such cases so that
the bounds are changed on the WCT/transition that is about to apply. It
also keeps track of these transition tokens and listens for
onTransitionReady() to verify when these immersive exits have completed
and thus the DesktopRepository state can be updated.

Flag: com.android.window.flags.enable_fully_immersive_in_desktop
Bug: 369443870
Bug: 369444146
Test: open YT music in immersive, minimize, unminimize - check bounds
are maximized but not immersive, repeat for other scenarios

Change-Id: I6c525433c8ac673a738f597d0cd0b344993a7ce9

Former-commit-id: 0a7aa27590daaa5fdd8d20a7de1ee542a2db026a