Refactoring gesture recognition logic
1/n of introducing gesture live progress tracking

Main motivation for refactoring is making gesture recognition more composition-friendly and extendable, currently ThreeFingerDistanceBasedGestureMonitor is counter-example of that. Goal is to easily filter by number of gestures and react on distance changes and - in the future - to velocity at the end. And to be able to have any combination of that.
Also I think the new code is easier to understand because we avoid inheritance.

Incoming further changes:
- Rename *GestureMonitor to *GestureRecognizer
- Refactor Recent apps gesture logic to follow similar design of compositing features

Bug: 369817369
Test: All tests are passing
Flag: com.android.systemui.shared.new_touchpad_gestures_tutorial
Change-Id: I690b8a95d93a4a18f41ade98e4b214d14153cd6a

Former-commit-id: 8a22cd588b69b98eb2e13d75e0faa92defd7b3a3