NotifCollection.dismissNotifications will now remove hidden summaries.
Previously we would take the list of notifications exactly, but that left room that a notification's summary, if it was not included in the list, would be left in the shade.  We now check if each entry is the sole logic child of a single summary, and if so we include that summary in the dismissal (assuming it was not already included), and generate the necessary stats object.

Bug: 355967751
Flag: com.android.systemui.notifications_dismiss_pruned_summaries
Test: atest NotifCollectionTest
Change-Id: Id3eda2f7a36227e4d5a921888735dd898d33a61a

Former-commit-id: 0248b5c81e77966bb8ffbac498a60f2bfd6c00c5