Set caption bar height to status bar height when in fullscreen
We make the assumption that the caption bar height is equal to the
status bar height. This is causing tests to fail for the cf tablet
target since this is not the case. This change assures that's the fact
by directly setting caption height to status bar height.

Bug: 372502136
Test: atest CtsWindowManagerDeviceAnimations:android.server.wm.animations.BlurTests#testAllBlurRemovedAndRestoredWhenToggleBlurDisabled
Flag: EXEMPT bugfix
Change-Id: I00b7bfc391bc89e715002fa3f8dbf00e9f03b9f4

Former-commit-id: 426870c0d2c2bca472585ce76ed8cb8d4cae9e86