Fix persisting SFPS indicator issue.
KeyguardBouncerRepository#alternateBouncerVisible state is incorrectly set and then never reset.

This is caused by usages of show()/forceShow() that only read (but not collect) the canShowAlternateBouncer flow to set the alternate bouncer visible state, if there is a race condition between when canShowAlternateBouncer flow changes to false and when the read happens, alternateBouncerVisible state will be set to an incorrect value and not reset on time.

Confirmed with additional logs that this is what is happening.

Flag: EXEMPT bugfix
Fixes: 360737693
Test: verified manually
 1. Enable screensaver while charging (on a large screen device with side fingerprint sensor)
 2. Go to screensaver
 3. Try authenticating with un-enrolled finger 5 times
 4. Bouncer should be shown
 5. SFPS indicator should not be visible on bouncer
 6. unlock the device with PIN/pattern/password
 7. SFPS indicator should not be visible after device is entered
Change-Id: I3bad4783b4283820bcf6473fd87da8cc8173c07f

Former-commit-id: 070817baf2efe665c186e19775c9beda05199a72