Prevent premature-ready between activity finish and completePause
when starting an activity, due to lifecycle mechanics, WM waits
for the prior activity to pause before starting the new activity.

If an activity starts a new activity and then finishes itself
immediately, there is a short-window where neither it nor the
next activity is visibleRequested so if a syncgroup is already
marked ready, its possible for the sync to finish prematurely.

Normally, we try to manage the readiness; however, if the
activity-start came from WCT, the SyncGroup will be marked
allready.

This CL adds a check to prevent a pausing activity from being
considered sync-finished until visibility is recalculated in
completePause

Bug: 361127213
Test: Launch a trampolining task (eg. calendar) from a shortcut
      (specifically via WCT pendingIntent) multiple times and
      observe no Transition error message.
Flag: EXEMPT bugfig
Change-Id: Ie7733a62eea865f1f67f20ae5cc162c57325e190

Former-commit-id: c90595c9e6b5c305176a76b2212ba30c7e2f0fdd