Default to hardware vibration attribute for SCROLL_* feedback
In ag/29002964, we changed the default vibration attribute used for the SCROLL_* constants from hardware to touch. Although this is our goal in the long term, it's causing a bug (b/372380778), because there are older usages of the View.performHapticFeedback API that directly play the SCROLL_* constant without providing the input source. We will fix this in the long term via b/372820923.

Bug: 372380778
Change-Id: I34c1b08bba1a0752ee190dd904b2d60395f98335
Test: atest HapticFeedbackVibrationProviderTest
Flag: android.os.vibrator.haptic_feedback_input_source_customization_enabled

Former-commit-id: be573449742a40bd6c8fa3d063bc66fc416dd222