Calculate Occulusion base on TDA status.
We currently have a single Flag that globally sets the entire keyguard
state as occluded vs not occluded. On top of this, we can pass TDA
status and base on whether it's moving to the front or to the back, we
can figure out occlusion on a per-display level.

Bug: 364930619
Test: atest android.server.wm.keyguard.KeyguardTests
Flag: EXEMPT refactor

Change-Id: Ie943e7ef6fc0648f38b0883e66eb9c0f6825d0f8

Former-commit-id: 26e46b53addd865450d4e3d4d5e3f1ff1090f34d