Use `audioSharingInteractor.audioSharingAvailable` instead of calling BluetoothUtils.
Also fix a minor bug: previously the dialog only listen to `onAclConnectionStateChanged` when a device is disconnecting as we assume that `onProfileConnectionStateChanged` must be triggered right after if the device is connecting, however this is not true for BLE devices as they don't have profiles. We should remove that restriction.

Test: atest
Bug: b/360759048 b/358276696
Flag: com.android.settingslib.flags.enable_le_audio_sharing
Change-Id: Ibc6c9f3801722daa1be5556ded18061c64d0ddfb

Former-commit-id: a67c924de154782bb7f0ec5e283411caac4385e8