Consider SAW permission for real caller
This commit addresses an issue where the BAL (background activity launch) logic was not properly considering the SAW (System Alert Window) permission for the real caller.

Added logic to check for the SAW permission when determining the exemption in the path for the real caller (analog to the caller).
Added a test to verify the fix.

For more context, see the Android documentation on background starts: https://developer.android.com/guide/components/activities/background-starts

Test: atest BackgroundActivityStartControllerExemptionTests
Flag: com.android.window.flags.bal_additional_start_modes
Bug: 327738445
Change-Id: I0594d47d15d9bc9aaf7b7da9242000c61686403b

Former-commit-id: 65f452b596237004cdba760b6164988452e90314