Use correct navigation hint for IME visibility
The system navigation bar and taskbar were correctly computing the
navigation hint flags, but using the incorrect flag
(navigation_hint_back_alt) to save the state of the IME visibility.

While this back_alt flag can only ever be set while the IME is visible,
it can also be unset when the IME is visible. In practice, any IME can
override this to be unset while the IME is visible, and the system also
temporarily overrides it while the IME Switcher menu is visible. This
would lead to the navigation bar/taskbar incorrectly beliving the IME is
not visible.

Additionally renames onImeVisibilityChanged (introduced in [1]) to
onBackAltChanged, to reflect the original meaning from [2].

This fixes the IME visibility state to be computed based on the
navigation_hint_ime_shown flag.

  [1]: I6034d66532d83d7f12b5e7d65610d4aa91b3533f
  [2]: Ifd1f8c9f400d90542f0ca858b9a4deacabbd518a

Flag: EXEMPT bugfix
Bug: 366129400
Test: atest NavigationBarTest#testSetImeWindowStatusSysuiState_ImeVisibleImeSwitcherButtonVisible
  NavigationBarTest#testSetImeWindowStatusSysuiState_ImeVisibleImeSwitcherButtonNotVisible
  NavigationBarTest#testSetImeWindowStatusSysuiState_ImeNotVisibleImeSwitcherButtonVisible
  NavigationBarTest#testSetImeWindowStatusSysuiState_ImeVisibleBackDispositionAdjustNothing
Change-Id: Ic57cea49f5ff49132802083b4f0c9b0e82db1cf7

Former-commit-id: 1d0b620cea46d3f154428d73c6489e235fe83439