Do not handle touches next to shelf
NSSL would return true for all touches below the last notification,
and to the right of the shelf. This is essentially empty space.
The reason for this is the SwipeHelper is always returning true,
which would set horizontalSwipeWantsIt == true even though that
is clearly not true on a tap.

This causes an issue with tapping the unlocked icon when the
last notification is directly above it. The shelf was taking
the touch directly over lock icon, preventing device unlock.

Instead, use the swipe information from the intercept event which
seems to be more accurate, and carry this state to the onTouch
event.

Fixes: 370270324
Fixes: 358424256
Test: manual - horizontal swipe notifications
Test: manual - vertically swipe notifications
Test: manual - all touch interactions with notifs
Test: manual - tap on shelf to expand shade
Flag: EXEMPT bugfix
Change-Id: I906da1cf5e5106ada355232941399efe7f1d3b14

Former-commit-id: 0b558a935f18be75b71f06b81d8a807c3fca4c3d