Unify current session in vibrator service
Update vibrator manager service to have a single current session that
can be either an external vibration or a vibration effect, in
preparation to add support for ongoing vendor vibration sessions.

The update preserves the existing behaviour, so some of the listeners
for settings updates and system broadcasts are explictly ignoring
ongoing external vibrations. New bugs were open to investigate and
update that behaviour.

Bug: 345414356
Flag: EXEMPT refactor
Test: FrameworksVibratorServicesTests
Change-Id: I08ebe4049e79b8148539f4fb8af943a5c65e6e48

Former-commit-id: 8e1c638b0c80690874f9d3c3af9b257776b1469d