Remove inconsistent @Deprecated behavior in *removed.txt files
Previously, `@Deprecated` was handled differently in `*removed.txt`
files compared to `*current.txt`. In the former `@Deprecated` on a
class was not propagated to its class members or nested classes but in
the latter it was. This change removes that inconsistency by making it
behave the same way in the `*removed.txt` files as it already does in
the `*current.txt` files.

Bug: 373584764
Test: m checkapi
Change-Id: I828c94e7c7094bae2a619c9a4367688fce1b8c61

Former-commit-id: a571f1c927f3b0ca6febb1149a3e7c56910dd83e