Check event in app's bounds using absolute coordinates.
Event's X and Y coordinates are relative to the current window,
not necessarily the app, meaning that windows with different
coordinate systems than the apps - for example dialogs -
would have incorrect calculations if used. Also, even if the tap
is outside the current window's bounds, but inside app's bounds,
it should be forwarded to the app (for example to trigger dialog
dismiss).

Flag: com.android.window.flags.scrolling_from_letterbox
Test: atest FrameworksCoreTests:LetterboxScrollProcessorTest
Test: verified dialogs are dismissable.
Fixes: 372918146
Change-Id: I01f57aacf8b2aa04854f76c0d5cb0417b797a8aa

Former-commit-id: 44cb19410102507f39e4f0bda863b96c87ddf9f7