Add API for Custom IME Switcher button visibility
This is a follow up to the IME Switcher revamp [1] and the ability to
to request hiding the IME navigation bar [2].

The visibility of the IME Switcher button is decided by the system. If
the button should be visible in the IME navigation bar, but this bar was
requested hidden by the IME itself, then it is the responsibility of the
IME to draw an equivalent Custom IME Switcher button.

This adds a new API for IMEs to be notified when the requested
visibility of a Custom IME Switcher button changes.

  [1]: I1005cb6b10682f3e91a7ed847c290fcc206b5faa
  [2]: I8793db69fb846046300d5a56b3b0060138ef4cd5

Flag: android.view.inputmethod.ime_switcher_revamp_api
Bug: 358031444
Test: atest InputMethodServiceTest#testOnCustomImeSwitcherButtonRequestedVisible_gestureNav
  InputMethodServiceTest#testOnCustomImeSwitcherButtonRequestedVisible_threeButtonNav
Change-Id: I243c8d3d76bc412c6498e255dacca31f4b28863e

Former-commit-id: 807d6e72b055fef37aaf89d1bbb857d45b2a96e1