Defer resume of activity while applying wct
When the hierarchy is changing, defer resuming the activity until it is
ready to let it finish processing the wct.

If the activity does not wait, there will be extra lifecycle states even
if they are applied in the same wct. Update test to not defer resume at
the start.

Flag: EXEMPT (bug fix)
Bug: 369186876
Test: atest CtsWindowManagerDeviceActivity com.android.server.wm.ActivityTaskManagerServiceTests
Change-Id: I2f564f4b7245de94871999a4da122489dba7fd7d

Former-commit-id: 724a6ecd19ea7e266450d245e192eee51f2aa9eb