Implement NDK createSessionUsingConfig API
This patch implements a new way to create ADPF hint session,
with a new pacelable SessionCreationConfig. This object encapsulates
the required information for session creation, with a set of setters to
assign values to those fields in the session creation config.

This object can be reused on the client side in order to streamline
the session creation process.

Bug: 362801903
Bug: 367803904
Test: atest HintManagerServiceTest
      atest PerformanceHintNativeTest
Flag: EXEMPT_NDK
Change-Id: I12baa7064573a8281f034c1e0280d06947d97f30

Former-commit-id: dffd67c90715716972a57cbaf3badaa140cb3ad5