Fix flickering of the resize frame when resizing
After resizing, the frame flickers briefly due to the offsets not being
immediately set to 0 after resize. To remedy this, we just key the
viewmodel on the current size - and when the current size changes, we
create an entirely new viewmodel, which removes any existing offsets.

Fixes: 372230926
Test: manually by expanding/shrinking widgets and verifying it no longer
flickers
Flag: com.android.systemui.communal_widget_resizing

Change-Id: Iac45e9233d7be9dde4b8dabfd69df7f19c99d9cf

Former-commit-id: 3b4410086fbda2f31a04351936ce8d6552b8a962