Route dream touch handling to scene container
We want to reuse the existing dream touch handling logic with the scene
container, so route touches there from the various touch handlers as
needed.

There are still a few issues with the interactions themselves, but
verified the touches are sent and received as expected.

Bug: 371340996
Fixed: 371340996
Test: atest BouncerFullscreenSwipeTouchHandlerTest BouncerSwipeTouchHandlerTest ShadeTouchHandlerTest
Flag: com.android.systemui.scene_container
Change-Id: I20630441584e67caa18afb166e4059f386687a37

Former-commit-id: fb1033cbb8eba35495f80d3a2be4f95b1c5d0627