Topology get and set API
Move DisplayTopology to a package where it is visible to system apps. Make it implement Parcelable so that it can be passed between processes.

Create a fake permission enforcer in DisplayManagerServiceTest so that we can grant the MANAGE_DISPLAYS permission where needed. This needs a new constructor for BinderService so that it uses the mocked context.

Bug: 364907048
Flag: com.android.server.display.feature.flags.display_topology
Test: DisplayTopologyTest, DisplayCoordinatorTest, DisplayManagerServiceTest
Change-Id: Iafa68d6fbecf7860bd4a1f7a167dcc021093a698

Former-commit-id: 7853e4a536a99829a231c407f0aa425ba6f8542b