[Forensic] Add DataAggregator
The DataAggregator is in charge of aggregating the data from
the list of DataSources.
Whenever new data needs to be tranferred from DataSource,
the data go through DataAggregator and ForensicService.
The DataAggregator can also be used to enable and disable
the data collection in the list of DataSources.

Bug: 365994454
Test: atest ForensicServiceTest
Flag: android.security.afl_api
Ignore-AOSP-First: security feature
Change-Id: I5363dff2b2ca0b5b1631ba431a858b5c6ec9f786

Former-commit-id: 54708ada8a86f93ef6563dac5d84281e861e3dae