Add immersive option to maximize menu
Shown when the task is requesting immersive.

Also changes the maximize drawable to have a small bottom padding so
that it is slightly different than the immersive drawable. This change
also hides the snap options when the app is not resizable. Overall, this
means that the menu can now have between 1-3 options in it, depending on
resizeability and immersive requests, so the menu width and position are
adjusted accordingly.

Flag: com.android.window.flags.enable_fully_immersive_in_desktop
Fix: 369441303
Test: manual - open maximize menu on immersive apps, non immersive apps,
resizable apps and non resizable apps and check the options are shown
accordingly
Test: atest DesktopModeWindowDecorationTests
DesktopModeWindowDecorViewModelTests

Change-Id: Ie8c7357e55d9650cc318de5f6fdc791de0629db1

Former-commit-id: 6439678d821523cc8733434ab74a2c24207d5ac6