Adjust the process group of the phantom processes as well
When an app process is assigned to a new process group, do the same for
all phantom processes created by the app. Otherwise, those phantom
processes can remain in the top-app cpuset even after the app process
goes to foreground or even background.

Bug: 375058190
Test: run Linux terminal, and then go to launcher. check the below
3:cpuset:/foreground
2:cpu:/
1:blkio:/
0::/uid_1010291/pid_9233

ensure that the cpuset is in foreground, not in top-app.

Flag: EXEMPT bugfix

Change-Id: I7115ce86d3363e83b1808fb2db5bd2c143bed885

Former-commit-id: b244fdfa903df89cc0ada32d5165c2561a3f002e