Show privacy dot on connected displays
Introduces PrivacyDotWindowController, which adds the privacy dots
to WindowManager.

This controller is only used for connected displays for now, and the
old logic in ScreenDecorations is used for the default/main display.

Also extracts the ScreenDecorations ui thread to a Dagger module, so
it can be reused in PrivacyDotWindowController.

Test: PrivacyDotWindowControllerTest
Test: PrivacyDotWindowControllerStoreImplTest
Test: PrivacyDotViewControllerTest
Test: DisplayWindowPropertiesRepositoryImplTest
Test: Add a connected display and start a voice recording. Also change
      rotation of the display and see that the dot positions correctly.
Bug: 362720432
Flag: com.android.systemui.status_bar_connected_displays
Change-Id: I5f0b8ddb37b55a53b66f3357bc1279e4b765e744

Former-commit-id: 965caedb30bbb9bd954dbc1efe857c114fc6e359