Hide last typed number of pin by default on devices with a keyboard
When entering a pin on the lockscreen using the on-screen or hardware
keyboard (built-in or connected through other means) the number was
visible by default for some time. This change changes the default value
on devices that have a hardware keyboard to replace the number with a
symbol immediately.

Flag: com.android.internal.widget.flags.hide_last_char_with_physical_input
Test: atest LockPatternUtilsTest
Bug: 339270220

Change-Id: I2c9e38f77ff8f849729815e0ec136aae2b6062c5

Former-commit-id: a85579687d7e60487ed123cdc58a8ee7e4f228f0