Add WindowManager flag for capturing the power key event in the foreground.
This is part of a feature to grant Wallet access to the double-click
power button gesture while the app is in the foreground.

Wallet would need to set this flag for specific windows it wishes to
handle the power gesture.

Bug: 357144512
Test: atest WmTests:WindowManagerServiceTests
Flag: com.android.hardware.input.override_power_key_behavior_in_focused_window

Change-Id: I03c97cac407afca81ed193a33524717e0aed92f0

Former-commit-id: 2d5b6806e3a184e358bcd4eb873a199ccfa8fdfd