Exclude the PIC key package_info from shared memory
The PIC key package_info is being sniffed as a sysprop outside of the
cache infrastructure, so it cannot be moved into shared memory yet.
This change excludes that one key from shared memory (when the feature
flag is enabled).

This is expected to be temporary until a mechanism is provided to
clients to sniff nonce changes in controlled way.

This change also corrects the bug numbers in the trunk stable flags
related to this effort.

Tested by booting an image with the flag enabled and verifying that
the package_info cache key exists as a system property, but other
system_server cache keys are not properties.

Flag: android.app.pic_uses_shared_memory
Bug: 360897450
Test: atest
 * FrameworksCoreTests:PropertyInvalidatedCacheTests
 * FrameworksCoreTests:IpcDataCacheTest
 * CtsOsTestCases:IpcDataCacheTest
 * CtsAppOpsTestCases
 * CtsMediaAudioTestCases
 * CtsVirtualDevicesAudioTestCases
Change-Id: I0250beac5b8e1d85c1a2ff21de8408f07f074a45

Former-commit-id: 77c12e0fbd8d7f1eb0976a52f7d8f4849fb87d44