Add aspect ratio setting button in window handle menu
The button will be visible if the task is fullscreen windowing mode
and the display is ignore-orientation-request (i.e. large screen).
Currently it only applies if desktop mode is enabled.

This is to make it easier for users to switch aspect ratio of app.

Bug: 376395315
Flag: EXEMPT add a common button
Test: adb shell setprop \
    persist.wm.debug.desktop_mode_enforce_device_restrictions false
  adb shell settings put global enable_freeform_support 1
  adb shell settings put global override_desktop_mode_features 1
  adb shell cmd window set-ignore-orientation-request 1
  adb reboot
  Open an app in fullscreen mode. Click top bar to see the button.

Change-Id: I5908286624cd426daeca317e9803e6619fe994be

Former-commit-id: dde9ccdcd56e1e50273f2b07f820798f313bc2ce